requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb 
sentence_transformers 
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
langchain_core
langchain_ollama
sentence-transformers
pandas
tf-keras